customized style, fixed errors in console caused by dialog components
